DB Clear

+++ bfwnetzwerk/views.py
@@ -44,6 +44,7 @@ def upload_file_prod(request):
 
 
 def read_csv_reha(f):
+    clear_db = True ## Beim Einlesen vorher alle Datenbanken löschen
     # Felder CSV
     feld_massnahmentitel = 0
     feld_massnahmeart = 1
@@ -62,14 +63,14 @@ def read_csv_reha(f):
 
     # Alle Datensätze löschen
     # Datenbanken leeren
-
-    Produkt.objects.all().delete()
-    Massnahmeart.objects.all().delete()
-    Organisation.objects.all().delete()
-    Schlagwort.objects.all().delete()
-    Abrechnungsart.objects.all().delete()
-    Kostentraeger.objects.all().delete()
-    Abschluss.objects.all().delete()
+    if clear_db:
+        Produkt.objects.all().delete()
+        Massnahmeart.objects.all().delete()
+        Organisation.objects.all().delete()
+        Schlagwort.objects.all().delete()
+        Abrechnungsart.objects.all().delete()
+        Kostentraeger.objects.all().delete()
+        Abschluss.objects.all().delete()
 
     with open(f, encoding='utf-8') as csvdatei:
         csv_reader_object = csv.reader(csvdatei, delimiter=';')
